parse: use *Lit for heredoc bodies
This ends the whole Redirect pointer madness. Some of its issues had
been fixed, but there were more to uncover. Presumably because nested
binary expressions would still copy the structs from one place to
another, meaning that our previous fix was not sufficient.

This should fix it for good, as we no longer depend on where Stmt is
present in memory.

Fixes #7.