shfmt: use buffered writer for files
If -w is used but -l is not, we were writing directly to the file. This
is apparently a bad idea, as it results in tons of unbuffered writes.

Below is an example on a very large set of files.

Before:

	real    0m10.747s
	user    0m6.347s
	sys     0m4.420s

After:

	real    0m5.961s
	user    0m5.650s
	sys     0m0.280s

This isn't necessary for stdout, as it is already buffered.