printer: fix regression in 176d402a
We can look at performance again once these tests are in place.

Don't revert the Lit part - see the tests - since it would be far from
trivial. We would essentially have to keep track of the end positions of
literals.

Instead of all that extra work, for now simply keep the current
behaviour of adding lines as seen in the tests.