all: re-use ast.ArithmExp for conditions
Back when I added support for "if ((foo))", I did not know that ((foo))
was actually a valid command. Drop all this ast.Cond interface
abstraction, as it's just another command.

Drastically reduces complexity in conditionals. The only added nit is
that ast.ArithmExp is now both a valid command and word part - if and
only if started with (( and $((, respectively.