parser: backtrack to retry $(( as $( (
Turns out that a newer POSIX spec does require backtracking, since it
introduces ambiguity in $((. If what follows $(( is not a valid
arithmetic expression and wasn't properly closed with )), retry parsing
it as a subshell in a command substitution, i.e. as if it were $((.

The diff is rather large because to see if the terminating )) is there,
we have to tokenize all the way to it. And since we're not doing lazy
tokenization - we're checking whether arithmetic operators are valid
right away, for example - we must delay the errors until we know what to
do with them.

See the TODO as this is probably not complete yet. That possible
breakage would only affect programs that use a weird combination of
ambiguous $(( and nested statements inside them though, so it should be
safe to leave that to be fixed later.

Fixes #30.