parser: use state type instead of tokens
This makes the behaviour of the parser and tokenizer much more
descriptive. Until now, we were re-using tokens as a way to lazily
describe what state we were in.

This was handy since we were sometimes re-using those values to decide
whether or not we should stop upon finding certain tokens. The code is
more verbose now since that has to be checked explicitly, which is good
since it improves readability and doesn't hide away what's going on.

No performance changes.