parser: speed up assignment parsing
Deciding whether or not a literal token is or not an assignment isn't
trivial. We were doing it in getAssign(), which was called on each
leading call literal until an argument was found.

Both in cases where there was an assignment and there wasn't, this was
costly. Not only were we calling this function, but we were also calling
strings.Index and doing some extra logic.

The tokenizer already iterates over the literal's bytes, so this work
can be done in advance there. Sometimes that work will be redundant,
like for arguments that cannot be assignments. But the byte '=' is not
that common anyway, and the speedup is noticeable.

name                    old time/op    new time/op    delta
Parse/Cmds+Nested-4       8.31µs ± 1%    8.17µs ± 0%  -1.67%          (p=0.002 n=6+6)
Parse/Assign+Clauses-4    5.82µs ± 0%    5.69µs ± 1%  -2.11%          (p=0.002 n=6+6)