parser: don't slice past input length
In a few extreme cases involving a backslash at the end of the input, we
would incorrectly slice past the input byte slice. This didn't crash
because of memory alignment, and didn't matter because this would only
happen in errors where the values are discarded.

Make a couple of tests actually trigger this by having the byte lengths
be 8, thus making len == cap.