parser: take a safer approach for $(( ambiguity
Up until now, we were fully parsing the bytes twice. This was
troublesome because lots of things could go wrong and the parser could
get stuck if we entered nested states like quotes or expansions.

The POSIX spec says, about determining whether or not $(( if followed by
a potential arithmetic expression:

	The shell need not evaluate nested expansions when performing
	this determination.

So instead of fully parsing, simply tokenize without changing state and
see if we get to a )) at any point.

The spec is vague as to whether this is the correct thing to do:

	... the shell shall first determine whether it can parse the
	expansion as an arithmetic expansion and shall only parse the
	expansion as a command substitution if it determines that it
	cannot parse the expansion as an arithmetic expansion.

Our current tests pass with the new method, including yet another input
that was making the old method hang.

This also means that we go back to reporting the missing )) instead of
the left-most error in the arithmetic expression. This is acceptable
given that $(( is ambiguous and its behaviour depends on whether or not
)) is present.