parser: don't execute heredocs in nested contexts
A heredoc executes (reads its body) when the next newline is found. Or
so I thought. Turns out that only if the newline is not in a nested
context. For example, the following program is valid:

	<<EOF $(
		foo
	)
	bar
	EOF

I would have assumed it would error, as it would read "<<EOF $(" and
then the rest as the heredoc budy, thus erroring on the missing ). But
only the last two lines are the heredoc body - the command substitution
is grabbed correctly by both bash and dash.

Now, on top of keeping track of what heredocs we have pending to read,
we also have to keep track of which ones are from this very context.