CI: run go test with -short and without -v
We have so many tests that Travis seems to have issues displaying all of
the output with the RUN and PASS lines. It often gets truncated, which
isn't useful at all. The output is also too verbose anyway.

Also use -short since the version of bash on Travis is too old. A
previous version of Bash did not recognise [[ -R foo ]] properly, so it
was erroring:

	parser_test.go:106: Unexpected error in `bash -n` of "[[ -R a ]]": bash: line 1: conditional binary operator expected
		bash: line 1: syntax error near `a'
		bash: line 1: `[[ -R a ]]'

Simply disable the bash confirmation runs (the only non-short test) on
Travis for now. In the future, we should disable those tests if the bash
version is too old. For now, this is good enough.