printer: fix comment alignment if adding newlines
If we're adding newlines, answering the question "does this statement
have an inline comment?" becomes more complex. If multiple statements
rae chained in a single line and a comment is at the end, only the last
statement has the comment.

This would fix itself after a second run, but that's not good enough.