all: don't differentiate $() and ``
There is no reason to use `` over $(). Every style guide and bash
tutorial out there tells you to always use $().

But what made me go forward with this is that the POSIX Standard itself
discourages the use of ``:

	Additionally, the backquoted syntax has historical restrictions
	on the contents of the embedded command. While the newer "$()"
	form can process any kind of valid embedded script, the
	backquoted form cannot handle some valid scripts that include
	backquotes.

	Because of these inconsistent behaviors, the backquoted variety
	of command substitution is not recommended for new applications
	that nest command substitutions or attempt to embed complex
	scripts.

It still defines it though, to keep compatibility with some old shells.
As such, the parser will continue to accept it. The difference is that
it will now generate the same AST as $(), thus shfmt will replace ``
with $().

Updates #34.