ast: fix ArithmExp.End() for $[]
We were always adding 2 to the end of the expression, ignoring that it
could be ] as well as )). This was spotted by the added checkSrc checks.

The parser no longer converts $(()) to $[] - it's now up to the printer.
This is because ast needs the information to know whether to add 1 or 2
bytes.