parser: rename package to syntax
We will be merging the parser, ast, token and printer packages into a
single package. Having ast and token separate is a bit silly, and shell
has very few tokens anyway. And the parser needs close use of these two
packages, for which we were using internal packages.

Instead, simplify everything by merging into a single package.

This is the first step - more to come in separate commits.