syntax: run some parent tests in parallel
Adding t.Parallel() to each of their subtests actually slows things
down, probably because there are so many of them and they take so little
that the overhead makes it a net loss. But making the parent tests run
in parallel does win some time - it reduces `go test -short` from ~130ms
to ~90ms on my machine.