syntax: get rid of defaultPos
Now that we verify positions against the source string, testing of
whether or not the Pos/End functions give up is not useful. Get rid of
the unexported var.

This also simplifies the tests and lets us do t.Parallel() in the other
nodes test. Brings down `go test -short` to ~70ms.