syntax: dedup token stringifying code
We redid this in the printer to speed it up somewhat.

Whatever the speedup, this is wrong. It bloats the code, overcomplicates
the printer and potentially introduces bugs and inconsistencies.

We should speed up the String() methods instead.

name      old time/op    new time/op    delta
Fprint-4    4.03µs ± 1%    4.34µs ± 1%  +7.64%  (p=0.002 n=6+6)