syntax: generate Token.String automatically
Use a modified version of golang.org/x/tools/cmd/stringer that gets the
string value from the inline comment of each constant, if there is one.

This wins back a good piece of the performance lost in the printer due
to the use of String() instead of token switches. But most importantly,
this further reduces the amount of duplicated logic that is maintained
manually.

name      old time/op    new time/op    delta
Fprint-4    4.31µs ± 0%    4.15µs ± 0%  -3.83%  (p=0.002 n=6+6)