syntax: simplify End() implementations
We have some extra logic to avoid End() being non-zero if Pos() is zero.
This is a bit silly, on one hand because it's not useful and on the
other because it makes End() calls noticeably slower.

Lift this restriction, simplify the implementation and remove the test
check.

Also add a Pos to ExtGlob as it was subtracting, thus if Pos() was 0
End() would be negative and overflow.

name                   old time/op    new time/op    delta
Fprint-4                 4.02µs ± 1%    3.93µs ± 0%  -2.16%  (p=0.002 n=6+6)