cmd/shfmt: don't read whole file to check shebang
We only need the first 32 bytes. This is usually OK, but if there is any
gigantic file without extension, we will be reading it all into memory
for no reason.

It's not uncommon for binaries to weigh megabytes, so this would be a
waste of time and resources in many cases.