cmd/shfmt: remove optional pprof
This is mostly just useful when profiling on master, so it's not very
appropriate for releases. Most importantly, it's a flag that might go
away and it noticeably increases binary size.

The size of shfmt with -ldflags='-s -w' goes from 2.0MB to 1.7MB on
linux/amd64 with this change.