syntax: rework handling of unicode literals
Keeping track of the bytes in a literal was messy as we didn't know
whether a literal was ascii or not to be able to take the shortcut and
append a single byte.

Rework this to be done in p.rune. The upside is that p.rune already
knows this, plus we don't have to use p.rbs. On the downside, we will
append an extra byte when finishing a literal which will need a bit of
extra work, and newLit is not inlineable because it's not a leaf.

Most listerals have more than one byte, so this should be a net win. And
the inlining issue should be fixed by the compiler at some point.

name                   old time/op    new time/op    delta
Parse/LongStrs-4         5.73µs ± 0%    5.20µs ± 1%  -9.30%  (p=0.002 n=6+6)
Parse/Cmds+Nested-4      7.59µs ± 1%    7.71µs ± 1%  +1.62%  (p=0.002 n=6+6)
Parse/Vars+Clauses-4     5.57µs ± 1%    5.61µs ± 0%  +0.66%  (p=0.030 n=6+5)
Parse/Binary+Redirs-4    7.23µs ± 1%    6.83µs ± 1%  -5.40%  (p=0.002 n=6+6)