syntax: get rid of $(( and (( backtracking
Supporting it is tricky. The method in place until this commit wasn't
too complicated and worked fine.

The issue comes now that we want to not read the whole input into
memory. If we backtrack, that means we may have to hold any number of
bytes to be able to go back and parse the bytes again.

Not only that, it also gets unnecessarily complicated to report some
errors correctly. Even with the logic we already had in place - which
resulted in slightly incorrect parse errors - utf8 errors were being
mistreated.

Instead of making this spaghetti code even more complicated and keeping
the backtracking, remove it altogether.

Thanks to Jilles Tjoelker for bringing up how shells like FreeBSD's do
not backtrack on purpose. It is presumed that most POSIX Shells such as
dash and ash also avoid backtracking support on purpose.

This is technically going against the latest POSIX Shell standard, which
as I understand does define the $(( ambiguity and mentions the rules
such as the precedence of $(( over $( (. At the same time, though, it
does mention that $(( should always be used to mean an arithmetic
expansion.

So if a script does not follow that rule, it is up to the user to fix it
before feeding it to shfmt. If a user follows POSIX and common sense,
shfmt should still work the same.

Updates #30.
Updates #36.