syntax: move comments out of binary statements
For a binary statement x, we had solved an issue with the comments by
moving the comments after x.Y. But this breaks if x.Y is complex, like
having a command subitution as shown in the new test.

Instead, move all of the comments before x.Y in front of all of x.

It seems like keeping comments in between x.X and x.Y is impossible if
we use the google-style binary operands:

	foo \
		# this breaks the escaped newline
		&& bar

This would work, though:

	foo &&
		# this works
		bar

However, it's confusing either way and should be avoided. This doesn't
seem reason enough to drop the backslashes.

Fixes #53.