cmd/shfmt: don't ignore println and f.Close errors
Especially if we are writing to files, don't ignore the file close
error. This does mean that we will try to close the file twice if we're
writing to it, but the second close is in the form of "defer f.Close()"
thus the "duplicate close" error will be ignored.

The alternative would be manually doing f.Close() before each return,
which would be too complicated.