syntax: don't make unrune index p.bs
The added test shows how even 8 bytes of padding after the current byte
is not enough. We can still end up accessing p.bs[-1].

We could add padding at the beginning with old bytes in p.fill, but that
would complicate things. Instead, hard-code the rune we want to fetch
from the past since we know it at compile time.