cmd/shfmt: add -version flag for reporting version
I'm working on getting `shfmt` added to the default preinstalled stuff on the
Travis CI VMs, and I'd like to be able to include the `shfmt` version in the VM
metadata.  I realize that having a hardcoded version is a bit of a pain, and I'd
typically implement this as an injected string of `git describe --always --tags`
or similar, but didn't want to introduce such a big change without first
proposing the flag.