syntax: parse = as == inside [[ tests
The two mean the same, and == is much clearer. = is only there for
consistency with `test`, which prefers = over == for POSIX
compatibility. But since [[ is bash-only, == is supported in both
arithmetic and test expressions, and = can be confused with assignment
(as in arithmetic expressions), == is preferred.

Don't remove the constant since that would break backwards
compatibility. The meaning/behaviour of TsAssgn and TsEqual was always
the same, so removing the use of the former should not break any
programs.