fileutil: require shebangs to end with whitespace
Strictly speaking, we would like the line to end with a newline. But we
would need to support \r and any spaces before those. Expect a
whitespace to keep things simple.

Make sure we don't require an immediate newline with a test that uses
bash args.

Also update the minimum file length to that of the shortest shebang plus
a whitespace, which is 10 instead of 8.