syntax: fix heredocs ending with an empty line
The lexer was buggy - it would ignore a newline if followed by another
newline, leading to weird issues like that one.

This was added to ignore tabs, but we should be peeking and only then
reading, instead of reading first. We don't want to read if it's not a
tab.

Fixes #70.