syntax: add Filename to Position
This should have been done by the start. For some reason, we added it to
ParseError instead, which is a bit silly.

This technically breaks backward compatibility, but only in the
following code:

	_ = ParseError{Filename: "foo"}

ParseError is a fairly new type and it shouldn't be declared outside of
syntax, so it's a safe enough breakage. Moreover, ParseError.Filename
will still work.

This is an important enough fix to warrant the above.