syntax: stabilize printer with comments in pipes
The old logic was to push all comments in between both sides of a binary
command (such as &&, || and |) to the very beginning, before the first
command.

This works OK, but makes the printer unstable when comments are
positioned in somewhat weird ways. Stability here means that whatever
the printer formats, it will stay the same if it's parsed again and
re-formatted.

Instead of pushing commits to the beginning, leave them after the
operand somehow. Since we use backslashes, for the moment that means
separate lines.

Fixes #83.