syntax: support backslashes inside backquotes
Backquote command substitutions don't treat backslashes normally all the
time. In particular, they are "ignored" if followed by another
backslash, a dollar or a backquote.

This does the first of them. The other two are left in a TODO for
another day because they're even more insane. In particular, it can mean
that the lexer has to "give up" on a literal and instead tokenize a
backquote, if it sees something like `foo \`bar\`` (the first escaped
backquote).

Fixes #85.