syntax: error on arith assigns to non-names too
In 5ee78266 we enforced that the assign operators must follow a word.
Enforce too the fact that the word must consist of a single literal and
it must be a valid name, like assigns in commands.

As before, the same rules apply to ++ and --.

And add TODOs to consider splitting these up in 2.0.