syntax: no quotes and expansions in arith exprs
Quoted strings and extensions won't be parsed as the user expects
because we can't parse their contents statically. They serve no purpose
in arithmetic expressions anyway, since "$a" should simply be "a" and
there is no reason to quote anything.

If a script really wants to construct an expression at run-time, they
should be clear about it with "eval".

We also don't need to check for len(w.Parts) == 1 in assignments.

And add another TODO for 2.0.