syntax: replace Word by Lit/ParamExp in ArithmExpr
See the TODO.

Also simplifies some code, removes a level of indirection and lets us
more easily deal with non-names being assigned to.

Also remove the structure TODOs, since the indirection to *Lit in
assigns is one level instead of two now. And it wouldn't just be a *Lit
all the time, like in arr[0]=3 - see #94.

Updates #82.