syntax: better mechanism for :+ and :- spaces
To avoid changing a program fro ${foo: +2} to ${foo:+2} (note that :+ is
a token here) we must add a space if '+' or '-' will follow ':'.

The former method worked okay, but it broke easily if the expression
wasn't a simple UnaryArithm. Add such a test and fix it.

This test also covers the two previous commits; '+' breaking a literal,
and the precedence of unary expressions.