syntax: error on unescaped $ literal
The only context in which it could make sense is a double-quoted string
that needs to use expansions too. For example:

	sed "s/$foo bar$/with/"

In that case, the second '$' is a literal and not a parameter expansion,
as it is not followed by a literal nor a character that makes its token
longer like '{' or '('.

The program should instead be written as:

	sed "s/$foo bar\$/with/"

In all other cases where expansions are not necessary, string literals
should be used instead, i.e. single quotes.

This gets rid of split literals without extra complexity and forces
programs to be saner and easier to read for both the human eye and the
parser. It also means that some possible bugs may be caught, like the
use of $[foo] in POSIX mode that will now error.