syntax: remove TODO for splititng of lits on [
It's not ideal, but we don't want to fully parse [] and {} expansions
either. These are just literals, and whether or not they are expansions
depends on what characters is in them. For example, per the posix spec,
a[b/]c cannot extend to "abc a/c" because it contains '/'.

"Joining" these literals would mean either gluing them together in the
parser, or teaching the lexer when to not break literals on '['. Both
approaches seem pretty complicated for something that isn't a terrible
problem to begin with.