syntax: merge Pos and Position types
This has several advantages. The most important one is that positions
are self-contained, so we no longer depend on File.Position(Pos).

Another advantage is that the printer is now much simpler, since it can
work with lines without having to depend on a File.

The new positions are bigger - 64 bits instead of 32 - but they still
contain the offset byte, and the line and column numbers.

Keep filenames out of them for now, as that would either mean an extra
pointer or an extra string. Usually, if you need the filename, you
already have the *File or you can pass the original filename around.

Overall, the parser is around 0-5% slower. This is fine given the gains
in usability and simplicity.