syntax: undo parser restriction on $ literals
This was done thinking mostly of shfmt, the formatter. In such a
scenario, the parser being strict is a good thing.

But in a broader sense, the parser should prefer compatibility with
POSIX and Bash than aligning with shfmt features.

For now, undo the restriction. A similar feature - such as rewriting the
input program to escape dollar signs - might reappear in a different
place, like the printer or shfmt.

Fixes #130.