interp: settle on echo -E being the default
Now we understand what -e/-E do, as it's the same as in $''.

Make -E the default instead of -e. The former is the default on bash and
ash, and the latter is the default on dash and mksh. There's no clear
spec around this, so go with bash as it's the most common.