interp: allow running Stmt and Command nodes too
As well as a File. Can be useful to run a very simple program while
skipping the verbosity of constructing an entire File.

I can't think of any other node that would be useful or sensible on its
own. These should be fine for now.

This is now possible thanks to positions no longer being tied to the
root File node.

Updates #119.