syntax: revert rename of ParamExp.Excl
I renamed it to Indirect thinking that it was the only meaning of the
operator. But it turns out that, on top of ${!foo} meaning a level of
indirection, ${!foo[@]} also means all the keys in an associative array,
as opposed to all the values.

So undo the rename, as this bool must represent both.