syntax: disallow ! in a pipe command
Following the POSIX spec and all the shells we test against. In other
words, disallow a ! right after a pipe operator, among other scenarios.

Also test that "! !" works (the first negates, the second is the
command). This is valid POSIX, as "!" is only a reserved word and thus
can be used like one can do "echo if".

However, dash and mksh are wrong here.