all: move to mvdan.cc/sh/...
The most important advantage is that the project relies less on the
provider that hosts the source code. If GitHub ever goes down or
anything else happens to it, a switch to another provider would be much
less painful.

And of course, the import paths are a little shorter and less verbose.

For #150.