cmd/gosh: support exit commands properly
To do this, a new interp.Runner method was needed - one that executed a
node, but didn't treat it as a whole program. Run did treat its argument
as an entire program, so it wasn't useful - all exit codes would mean a
Run error, even those not from an actual exit call.

Also support other types of errors, such as internal interpreter errors.