syntax: parse valid parameter names only
When parsing $foo. We stopped on certain characters, i.e. a blacklist.
Yet the POSIX spec says the opposite:

	If the parameter is not enclosed in braces, and is a name, the
	expansion shall use the longest valid name (see XBD Name)

In other words, we should use a whitelist of characters that are allowed
to advance the lexer. This means that we should stop on any symbol and
any non-ASCII letter, for example. Do it this way and add more tests.

This also means we can simplify quite a bit of the code dealing with
these in the ever growing advanceLitOther.