syntax: include comments in Walk, if any
There is no reason not to. They can be ignored at parse time, and in the
walk callback function one can simply ignore them when found.

When walking them, make sure they are walked in order by position. This
is including when comments can cover both the lines leading up to a node
and the line itself, as an inline comment.

Add a test for all these things, and also test that all nodes we expect
to see are seen - no more and no less.

Fixes #164.