syntax: remove need for a rune var in p.next
This was added for a noticeable performance improvement of around 3%
when parsing strings, due to how advancing ASCII runes is a very hot
path.

However, as it turns out, the compiler can realise this with a bit of
help - if we write p.r just before reading it. Then it can know that
it's safe to reuse the same value without the memory read.

Confirmed that the benchmarks do not change with the code
simplification.