syntax: fix comment padding edge case with {
Since that's still a comment in the following line but not really an
inline comment (since it doesn't follow the end of that statement), it
was throwing off the logic.

Instead of trying to support consecutive comments no matter what, make
these break the padding properly. Add a test too.