interp: treat PWD like a var, like HOME
That is, it is set at start-up and whenever cd is called. Otherwise, it
can be changed at any point. This is what bash does too.

While at it, I realised that the creation of subshells wasn't correct. A
shell may modify its vars map, and we have to make sure a subshell can't
modify the map of the parent. It could before, as the map pointers were
the same.