syntax: make the Stmts callback return bool
So that one may stop the parser prematurely if needed. Subsequent reads
may still happen if there is a following token, but the parser will not
continue any further.

Doing this change now, as the signature of Stmts will be fixed once we
release it in a version such as 2.1.