all: introduce KeepPadding and shfmt -kp
This option will let a user keep their own padding and alignment of most
nodes and tokens, while still taking advantage of the parser and printer
for everything else.

See the godoc added in KeepPadding for details, and the discussion in
the issue below as to why the feature was added and why it works the way
it does.

Fixes #168.