cmd/shfmt: add -f to list all shell files
Listing all the shell files in a repository can be hard, as there is
often a mix of different filename extensions and shebangs.

The tool already knows how to deal with this, but the functionality was
just internal and not available to any other tools or users on its own.

The new flag does exactly this. It is a bit wasteful - it does open all
the files that it finds, often unnecessarily - but let's not worry about
that for now, as the code is simple.

Fixes #154.