interp: clean up the read builtin code a bit
First, fix and get rid of the FIXME. We were not calling ifsUpdated if
IFS was updated in cmdVars.

Second, unindent a lot of the code with an early return.

While at it, clarify a few variable names, simplify an if/else bool
assignment, and replace a few "len(x) != 0" with "len(x) > 0".