interp: redo how ModuleExec uses $PATH
Before, it didn't use the interpreter's $PATH at all. With the
introduction of the interpreter's own $PATH, we made it so that we would
error if the file wasn't found before calling ModuleExec.

However, this is unfortunate. For example, if one wants to mock some
programs even if they are not installed, the change above makes it
impossible to do with ModuleExec, while it was possible to do before.

Change it so that it's DefaultExec that does the "executable not found"
error, letting the user do other things in their own module.

Fixes #194.