interp: fix concurrency bug in TestSignalSending
We were using the range variables in the concurrent goroutines (see the
t.Parallel call), which resulted in us testing only one of the test
cases multiple times.

With this fixed, the "ctx.KillTimeout <= 0" if body is now covered by
the tests. The bug was spotted because the tests never ran it before.