interp: move shell options into an array
And group flag and name strings in a single table, along with constant
indices to retrieve specific options without a linear search.

This has several advantages. For one, all the logic is in one place. It
will also be useful in the future when we need to list all the options
alphabetically, such as in the yet to be implemented "set -o".