interp: replace RunError with panics
RunError is not for regular interpreter errors, such as ExitCode. It's
for fatal, internal errors such as when an operator is unknown or when
see a type that we don't know.

Don't mix those with simpler errors, such as missing arguments on
builtins. Use regular exit codes for those, as we should.

Using panics for fatal errors is fine; one can still recover from them.
And it helps clarify how unexpected and unusual they should be.