syntax: add TranslatePattern and PatternRune
Moved from the interp package, now exported since they can be useful to
use shell patterns independently of the interpreter.

While at it, added an error return so that the function can now report
whether the input pattern was incorrect, and ensure that all returned
expressions are correct as long as the returned error is nil.

Also add unit tests, as before this was tested as part of interp's
end-to-end tests.

Updates #198.