syntax: add QuotePattern, remove PatternRune
QuotePattern is still as useful to the interp package, and it exposes
less internals to the user than PatternRune did. It's also very similar
to the regexp.QuoteMeta that some may already be used to.

More importantly, we were not escaping characters special to regexes in
TranslatePattern. Fix that too, and add a test.