syntax: deduplicate index expression parsing
Make it simpler and in one place. This will allow to tweak it soon
without having to worry about four different bits of code.

One error test is failing after this change, which will be fixed in the
commit right after.

This change also lets us remove the paramExpInd state, as we always use
arithmExprBrack now.