syntax: don't panic on TranslatePattern(`\`)
Assuming that backslashes were always followed by a character was an OK
assumption to make if the only uses were pattern matches in shell code,
as the parser already ensures that consistency. However, now that this
API is open, we should instead add a check and error if it's not the
case.

While at it, note how these funcs should not be used with file paths on
Windows.