syntax: error on parsing "! ! cmd"
In other words, forbid chaining of negation reserved words. Bash and
mksh allow this, but we can't allow it at the moment - that would
require breaking backwards compatibility in the AST. It also doesn't
seem very useful, and isn't a great idea as dash and other shells error
on it too.

While at it, add another weird reserved word test case.