cmd/shfmt: use a read-only open to check for shebangs
If a file is only readable but not writeable and does not
contain a valid shebang shfmt should not fail.

This implementation only opens the file twice if it
needs to write to it. If the file doesn't contain
a shebang or if no changes are necessary it is only
opened once.

Fixes #207.