syntax: improve parsing of regex expressions
In particular, stop parsing a literal out of them if we see a byte that
would make a token to be tokenized.

While at it, consider the '|' rune to not break literals in this state.
It is a valid part of a regular expression, so we shouldn't try to
tokenize it as a pipe.