syntax: fix backslashes in between backquotes
The way backslashes work inside backquotes is tricky - they are handled
when advancing runes, not when advancing tokens. Doing it at the level
of tokens is wrong, as there would be dozens of edge cases to take care
of when we think we'll see a literal (from the backslash) yet we could
see many other tokens instead.

This is unfortunate, as it means we must add an extra branch in the
quick path in the rune method. However, since backslashes aren't very
common, this shouldn't have a big effect on normal programs.