syntax: refine logic to skip backslashes in ``
Our old logic was to skip them outside of single quotes as long as they
were not followed by a newline. That is, we used a blacklist of chars
that it could be followed by.

But this is wrong. Quoting the spec:

	Within the backquoted style of command substitution, <backslash>
	shall retain its literal meaning, except when followed by: '$',
	'`', or <backslash>.

Use a whitelist instead, with those three characters. This also uncovers
a test case that we had written wrong. Fix it, after verifying with bash
that the newe version is the correct one.

Fixes #216.