syntax: initial support for nested backticks
That is, nested command substitutions that use backticks, such as:

	`echo \`foo\``

The lexer already supported backlashes within backticks, but it didn't
pass this information to the parser. Without that, the parser saw the
four backticks as equals, thus resulting in the second backtick closing
the first instead of opening a new nested command substitution.

Fixes #220.