syntax: allow quotes in arithmetic expressions
The reasoning for this, initially, was so that all arithmetic
expressions could be parsed statically. In theory, that makes for
simpler ASTs, and one could still use eval if running arbitrary programs
at run-time was necessary.

However, in practice that didn't work well. For one, the syntax package
isn't a great place to enforce this restriction. shfmt, a tool that
formats shell programs, would error on valid programs. And the error was
on purpose, when we could simply drop it.

But it also was a bad restriction. Even without quotes, one can still
construct an arithmetic expression at run-time:

	$ expr='3+4+5'
	$ echo $(($expr))
	12

For those two reasons, drop the restriction from the syntax package.
Packages that rely on syntax, such as interp, may need changes to adapt
to the new possible ASTs.

No programs depending on syntax should stop compiling or working
properly because of this change; all it does is stop erroring on some
inputs.

Fixes #223.