syntax: enforce indentation of <<- heredoc bodies
When we are indenting with tabs, indent the heredoc body and the closing
word too. Otherwise, remove all indentation to not mix spaces and tabs.

A side effect of this change is that <<- heredoc bodies in the AST never
contain the tab indentation as literals. This makes the AST simpler and
more accurate, as these leading tab characters are to be ignored when
interpreting the program.

Fixes #228.