syntax: fix printing of many stmts without pos
We were relying on them to have increasing line numbers. This is a
problem if the syntax tree was constructed manually, in which case it's
unlikely that it would have sensible position information.

The old code would print all statements on a single line without
semicolons. The new one forces newlines, as is expected when there are
many statements.