syntax: never parse a ParamExp with no parameter
We were doing it in the following case:

	echo $\
	 etc

Since an escaped newline can be part of an identifier, when the parser
sees "$\\", it cannot yet know if a correct parameter expansion will
follow. To work around the issue, if we end up with an empty literal,
fall back to a "$" literal instead.

Fixes #236.