syntax: make File.Pos and File.End use comments
When one wants to select a node such as a statement or a command,
comments should generally not be part of the selection. However, they
should be if the node is a file.

This is also more consistent with all other statement lists. All other
such lists are enclosed by tokens or reserved words, resulting in the
comments always being part of their Pos-End range. In this way, File was
inconsistent.

The change will also let us simplify the printer slightly, as it was
impossible to obtain the true start and end of the program.