syntax: simplify printing of statement lists
We were trying to be clever by avoiding work when we were printing zero
or one statements. This meant we could take some shortcuts, such as
avoiding the alignment of comments, or more easily deciding whether or
not a newline is required in each of the cases.

However, this added quite a bit of duplicated code. And, all in all, the
performance win was barely noticeable, as can be seen by the numbers
below.

Moreover, now Printer.stmts behaves in a more predictable way, so asking
it to use newlines is more straightforward.

name     old time/op    new time/op    delta
Print-4    6.42µs ± 2%    6.48µs ± 1%  +0.92%  (p=0.022 n=10+9)