syntax: add an option to space redirect ops
That is, to format them like:

	echo foo > file

This format is somewhat popular in the wild, and it is mentioned in some
prominent documents like 'man bash' and Classic Shell Scripting.

The option is very specific and doesn't clash with any other existing
options, so it shouldn't complicate the printer in the long run.

The Google Shell Style guide doesn't explicitly say how to format
redirections, so for now don't change its recommendation on the main
README file.

Don't change the default, as that would break backwards compatibility.
It's also not clear that this new format is vastly more popular than the
current one. For example, the POSIX Shell spec doesn't use spaces, and
the shfmt tool didn't have support for them.

Fixes #243.