interp: improve Runner.Dir tests
Use a temporary directory, to avoid issues when the test is run in
parallel with itself. Also add more cases, such as a path that isn't a
directory, and an invalid path that is absolute.

While at it, fix the code too. We only checked if the path pointed at a
valid directory if the path wasn't absolute.